Gluing it together. Converting fineoffset sensor. Tested!
The proofing of the new packet based pulse detection. The pulse
demodulator for PWM_RAW is added (with some minor corrections). The
fineoffset sensor is switched to the new decoder and tested to work
correctly (both with test data and live).
